pyproject.toml
[project]
name = "code-archeologist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "fastapi>=0.135.1",
    "groq>=1.0.0",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]
